page_analyzer/app.py
from dotenv import load_dotenv
from flask import Flask
import os


load_dotenv()
app = Flask(__name__)
app.config['SECRET_KEY'] = os.getenv('SECRET_KEY')


@app.route("/")
def hello_world():
    return "<p>Hello, World!</p>"
